+++ src/Utils.hpp
@@ -6,3 +6,5 @@ cocos2d::ccColor4F getTriggerCol(int id);
 
 std::vector<std::vector<GameObject*>> getClusters(const std::vector<GameObject*>& objs, float clusterSize);
 cocos2d::CCPoint getLineCut(cocos2d::CCPoint triggerPos, cocos2d::CCPoint p1, cocos2d::CCPoint p2);
+
+std::unordered_set<int> parseIntArray(const std::string& input);