+++ src/indicators/IndicatorVariables.hpp
@@ -20,9 +20,18 @@ namespace IndicatorVars {
     inline IndicatorType indicatorType = IndicatorType::Clustered;
     inline cocos2d::ccColor4F extrasCol1 = cocos2d::ccc4f(1.0f, 1.0f, 1.0f, 1.0f);
     inline cocos2d::ccColor4F extrasCol2 = cocos2d::ccc4f(0.0f, 0.0f, 0.0f, 1.0f);
+    inline std::unordered_set<int> groupBlacklist = {
+        1, 2, 3, 4, 5
+    };
+    inline std::unordered_set<int> triggerBlacklist = {
+        1, 2, 3, 4, 5
+    };
+    inline std::string settingsButtonTexture = "uwu";
 
     // the silly zone
     inline bool chroma = false;
     inline float sayoDeviceSensitivity = 0.25f;
     inline cocos2d::CCNodeRGBA* triggerIndicatorDrawChromaNode = nullptr;
+
+    void updateSettings();
 }